Add `CalibrationUseCase` to `AudioFeedbackHandler`.
