+++ app/src/test/java/com/moyi/liu/audiofeedback/domain/AudioFeedbackHandlerFrontBackTracksTest.kt
@@ -3,6 +3,8 @@ package com.moyi.liu.audiofeedback.domain
 import com.google.common.truth.Truth.assertThat
 import com.moyi.liu.audiofeedback.adapter.audio.StubAudioManager
 import com.moyi.liu.audiofeedback.adapter.transformer.SensorDataTransformer
+import com.moyi.liu.audiofeedback.domain.calibration.SensorCalibrator
+import com.moyi.liu.audiofeedback.domain.model.CalibrationConfig
 import com.moyi.liu.audiofeedback.domain.model.PowerAccumulatorConfig
 import com.moyi.liu.audiofeedback.domain.model.STUB_BOUNDARY
 import com.moyi.liu.audiofeedback.domain.power.StubPowerStore
@@ -10,7 +12,10 @@ import com.moyi.liu.audiofeedback.domain.power.getStubChargedIndicators
 import com.moyi.liu.audiofeedback.domain.sensor.SensorInitialisationFailedException
 import com.moyi.liu.audiofeedback.domain.sensor.SensorNotFoundException
 import com.moyi.liu.audiofeedback.domain.sensor.StubGravitySensor
+import com.moyi.liu.audiofeedback.domain.usecase.CalibrationUseCase
 import com.moyi.liu.audiofeedback.rx.StubDisposable
+import com.moyi.liu.audiofeedback.stub.StubMessageStore
+import com.moyi.liu.audiofeedback.stub.StubVoiceoverController
 import io.reactivex.rxjava3.android.plugins.RxAndroidPlugins
 import io.reactivex.rxjava3.core.Completable
 import io.reactivex.rxjava3.plugins.RxJavaPlugins
@@ -52,6 +57,12 @@ class AudioFeedbackHandlerFrontBackTracksTest {
     )
     private val powerStore = StubPowerStore(getStubChargedIndicators(2))
     private val audioManager = StubAudioManager()
+    private val calibrationUseCase = CalibrationUseCase(
+        messageStore = StubMessageStore(),
+        calibrator = SensorCalibrator(StubGravitySensor()),
+        voiceoverController = StubVoiceoverController(),
+        calibrationConfig = CalibrationConfig(2, 3)
+    )
 
     @Test
     fun givenSensorInitialisationFailed_handlerShouldThrowException() {
@@ -61,7 +72,7 @@ class AudioFeedbackHandlerFrontBackTracksTest {
             )
         }
 
-        AudioFeedbackHandler(sensor, audioManager)
+        AudioFeedbackHandler(sensor, audioManager, calibrationUseCase)
             .also {
                 it.powerStore = powerStore
                 it.dataTransformer = transformer
@@ -75,7 +86,7 @@ class AudioFeedbackHandlerFrontBackTracksTest {
     @Test
     fun givenDataStreamIsActive_whenCallingStart_shouldDisposeExistingOne() {
         val handler =
-            AudioFeedbackHandler(StubGravitySensor(), audioManager)
+            AudioFeedbackHandler(StubGravitySensor(), audioManager, calibrationUseCase)
                 .also {
                     it.powerStore = powerStore
                     it.dataTransformer = transformer
@@ -93,7 +104,7 @@ class AudioFeedbackHandlerFrontBackTracksTest {
         val sensor = object : StubGravitySensor() {
             override fun register(): Completable = Completable.error(SensorNotFoundException)
         }
-        val handler = AudioFeedbackHandler(sensor, audioManager)
+        val handler = AudioFeedbackHandler(sensor, audioManager, calibrationUseCase)
             .also {
                 it.powerStore = powerStore
                 it.dataTransformer = transformer
@@ -116,7 +127,7 @@ class AudioFeedbackHandlerFrontBackTracksTest {
     fun givenStartIsCalled_DataShouldBePassedToStream() {
         val sensor = StubGravitySensor()
 
-        AudioFeedbackHandler(sensor, audioManager)
+        AudioFeedbackHandler(sensor, audioManager, calibrationUseCase)
             .also {
                 it.powerStore = powerStore
                 it.dataTransformer = transformer
