+++ app/src/test/java/com/moyi/liu/audiofeedback/domain/AudioFeedbackHandlerLeftRightTracksTest.kt
@@ -5,11 +5,16 @@ import com.moyi.liu.audiofeedback.adapter.audio.StubAudioManager
 import com.moyi.liu.audiofeedback.adapter.transformer.MAX_GRAVITY_SENSOR_VALUE
 import com.moyi.liu.audiofeedback.adapter.transformer.MIN_SINGLE_NOTE_PLAY_INTERVAL_MILLIS
 import com.moyi.liu.audiofeedback.adapter.transformer.SensorDataTransformer
+import com.moyi.liu.audiofeedback.domain.calibration.SensorCalibrator
+import com.moyi.liu.audiofeedback.domain.model.CalibrationConfig
 import com.moyi.liu.audiofeedback.domain.model.PowerAccumulatorConfig
 import com.moyi.liu.audiofeedback.domain.model.STUB_BOUNDARY
 import com.moyi.liu.audiofeedback.domain.power.AFPowerAccumulator
 import com.moyi.liu.audiofeedback.domain.power.AFPowerStore
 import com.moyi.liu.audiofeedback.domain.sensor.StubGravitySensor
+import com.moyi.liu.audiofeedback.domain.usecase.CalibrationUseCase
+import com.moyi.liu.audiofeedback.stub.StubMessageStore
+import com.moyi.liu.audiofeedback.stub.StubVoiceoverController
 import io.reactivex.rxjava3.android.plugins.RxAndroidPlugins
 import io.reactivex.rxjava3.plugins.RxJavaPlugins
 import io.reactivex.rxjava3.schedulers.Schedulers
@@ -52,11 +57,19 @@ class AudioFeedbackHandlerLeftRightTracksTest {
     private val gravitySensor = StubGravitySensor()
     private val audioManager = StubAudioManager()
 
+    private val calibrationUseCase = CalibrationUseCase(
+        messageStore = StubMessageStore(),
+        calibrator = SensorCalibrator(StubGravitySensor()),
+        voiceoverController = StubVoiceoverController(),
+        calibrationConfig = CalibrationConfig(2, 3)
+    )
+
     @Test
     fun shouldReceiveLeftChargedSignal() {
         val handler = AudioFeedbackHandler(
             sensor = gravitySensor,
-            audioManager = audioManager
+            audioManager = audioManager,
+            calibrationUseCase = calibrationUseCase
         ).also {
             it.powerStore = powerStore
             it.dataTransformer = transformer
@@ -79,7 +92,8 @@ class AudioFeedbackHandlerLeftRightTracksTest {
     fun givenDirectionIsChanged_accumulatorsShouldBeEmptied() {
         val handler = AudioFeedbackHandler(
             sensor = gravitySensor,
-            audioManager = audioManager
+            audioManager = audioManager,
+            calibrationUseCase = calibrationUseCase
         ).also {
             it.powerStore = powerStore
             it.dataTransformer = transformer
