+++ app/src/test/java/com/moyi/liu/audiofeedback/domain/usecase/CalibrationUseCaseTest.kt
@@ -1,13 +1,14 @@
 package com.moyi.liu.audiofeedback.domain.usecase
 
 import com.google.common.truth.Truth.assertThat
-import com.moyi.liu.audiofeedback.domain.audio.VoiceoverController
 import com.moyi.liu.audiofeedback.domain.calibration.SensorCalibrator
-import com.moyi.liu.audiofeedback.domain.message.MessageStore
-import com.moyi.liu.audiofeedback.domain.model.*
+import com.moyi.liu.audiofeedback.domain.model.CalibrationConfig
+import com.moyi.liu.audiofeedback.domain.model.CalibrationResult
+import com.moyi.liu.audiofeedback.domain.model.Origin
 import com.moyi.liu.audiofeedback.domain.sensor.StubGravitySensor
+import com.moyi.liu.audiofeedback.stub.StubMessageStore
+import com.moyi.liu.audiofeedback.stub.StubVoiceoverController
 import io.reactivex.rxjava3.android.plugins.RxAndroidPlugins
-import io.reactivex.rxjava3.core.Completable
 import io.reactivex.rxjava3.plugins.RxJavaPlugins
 import io.reactivex.rxjava3.schedulers.TestScheduler
 import org.junit.AfterClass
@@ -89,44 +90,5 @@ class CalibrationUseCaseTest {
         )
 
         assertThat(voiceoverController.isDestroyCalled).isTrue()
-    }
-
-    private class StubMessageStore : MessageStore {
-        override fun getMessage(message: Message): String =
-            when (message) {
-                is CalibrationVoiceoverMessage.Preparation ->
-                    "preparation ${message.preparationTimeInSeconds}"
-                is CalibrationVoiceoverMessage.StartCalibration ->
-                    "start calibration ${message.calibrationDurationInSeconds}"
-            }
-    }
-
-    private class StubVoiceoverController : VoiceoverController {
-        val speakMessages = mutableListOf<String>()
-        var isDestroyCalled = false
-        var isInitialiseCalled = false
-
-        override fun initialise(): Completable = Completable.create {
-            isInitialiseCalled = true
-            it.onComplete()
-        }
-
-        override fun speakOut(message: String, type: VoiceoverController.SpeechType) {
-            speakMessages.add(message)
-        }
-
-        override fun speakWith(
-            message: String,
-            type: VoiceoverController.SpeechType,
-            timeoutMillis: Long
-        ): Completable = Completable.create {
-            speakMessages.add(message)
-            it.onComplete()
-        }
-
-        override fun destroy() {
-            isDestroyCalled = true
-        }
-
     }
 }
